make forgot password page a server component and create password recovery form to contain state
